closes #44

+++ server_modules/steganograph.js
@@ -11,111 +11,12 @@ router.use(function timeLog(req, res, next) {
     next();
 });
 
-router.post('/retrieve', function(req, res) {
 
+router.post('*',function(req,res,next){
 
-    var objectType;
-
-    // create an incoming form object
-    var form = new formidable.IncomingForm();
-
-    // specify that we want to allow the user to upload one file in a single request
-    form.multiples = false;
-
-    // store all uploads in the /uploads directory
-    form.uploadDir = path.resolve('./uploads');
-
-    var resDir = path.resolve('./results');
-    var fileCarrier;
-
-    // every time a file has been uploaded successfully,
-    // rename it to it's random hashed name
-    form.on('file', function(field, file) {
-        console.log('form field_TYPE= FILE field= ' + field);
-        var ext = '.' + file.name.split('.').pop();
-        fileCarrier = crypto.createHash('md5').update(Math.random().toString()).digest('hex');
-        fileCarrier = path.join(form.uploadDir, fileCarrier + ext);
-        fs.rename(file.path, fileCarrier);
-    });
-
-    form.on('field', function(field, data) {
-        console.log('form field_TYPE=FIELD field= ' + field);
-        if (field == "objectType") {
-            objectType = data;
-        }
-    });
-
-    // log any errors that occur
-    form.on('error', function(err) {
-        console.log('An error has occured: \n' + err);
-    });
-
-    // once all the files have been uploaded, send a response to the client
-    form.on('end', function() {
-        console.log('form end ');
-        var responseJSON = {};
-        if (fileCarrier.split('.').pop() != 'wav') {
-            responseJSON["status"] = "error";
-            responseJSON["error"] = "wrong data carrier extension";
-            responseJSON["error-id"] = "1";
-            res.json(responseJSON);
-            res.end('success');
-            return;
-        }
-
-        /* init as success. Will be changed if error appears*/
-        responseJSON["status"] = "success";
-        responseJSON["error-id"] = 0;
-        responseJSON["result"] = "";
-
-        var resExt = objectType == "file" ? ".file" : ".txt";
-        var resName = crypto.createHash('md5').update(Math.random().toString()).digest('hex') + resExt;
-        const ls = spawn('java', ['Java/src/steganography_tool/Steganography_Tool', 'Retrieve', objectType, fileCarrier, path.join(resDir, resName)]);
-
-        console.log('java process started');
-        ls.stdout.on('data', (data) => {
-            responseJSON["result"] += `${data}`;
-        });
-
-        ls.stderr.on('data', (data) => {
-            console.log(`stderr: ${data}`);
-            responseJSON["status"] = "error";
-            responseJSON["error-id"] = 2;
-            responseJSON["error"] = 'Processing error. Check files';
-        });
-
-
-        ls.on('close', (code) => {
-            console.log('java process ended');
-            /*remove uploaded files*/
-            fs.unlink(fileCarrier, (err) => {
-                if (err) {
-                    console.log("error deleting file carrier!");
-                }
-            });
-
-            if ((objectType == 'file') && (responseJSON["status"] == 'success')) {
-                responseJSON["result"] = resName;
-                /* Delete result file after 2 hours */
-                setTimeout(() => {
-                    fs.unlink(path.resolve('./results/' + resName));
-                }, 7.2e6);
-            }
-            res.json(responseJSON);
-            res.end('success');
-
-        });
-
-
-    });
-
-    // parse the incoming request containing the form data
-    form.parse(req);
-});
-
-router.post('/hide', function(req, res) {
     var objectType;
     var object;
+
     // create an incoming form object
     var form = new formidable.IncomingForm();
 
@@ -129,6 +30,7 @@ router.post('/hide', function(req, res) {
     var fileCarrier;
     var fileObject;
     var carrierExt;
+    
     // every time a file has been uploaded successfully,
     // rename it to it's orignal name
     form.on('file', function(field, file) {
@@ -138,13 +40,13 @@ router.post('/hide', function(req, res) {
             fileObject = crypto.createHash('md5').update(Math.random().toString()).digest('hex');
             fs.rename(file.path, path.join(form.uploadDir, fileObject + ext));
             object = path.join(form.uploadDir, fileObject + ext);
-        }
-        else {
+        } else if (field == "fileCarrier") {
             carrierExt = ext;
             fileCarrier = crypto.createHash('md5').update(Math.random().toString()).digest('hex');
             fileCarrier = path.join(form.uploadDir, fileCarrier + ext);
             fs.rename(file.path, fileCarrier);
-
+        } else { // Delete file 
+            fs.unlink(file.path);
         }
 
     });
@@ -177,16 +79,25 @@ router.post('/hide', function(req, res) {
             return;
         }
 
-        /* init as success. change if error */
+        // init as success. change if error 
         responseJSON["status"] = "success";
         responseJSON["error-id"] = 0;
+        responseJSON["result"] = "";
+        
+        // result file extension 
+        var resExt = (req.path == '/hide') ? carrierExt : (objectType == "file" ? ".file" : ".txt");
 
-        var resName = crypto.createHash('md5').update(Math.random().toString()).digest('hex') + carrierExt;
+        var resName = crypto.createHash('md5').update(Math.random().toString()).digest('hex') + resExt;
 
-        const ls = spawn('java', ['Java/src/steganography_tool/Steganography_Tool', 'Hide', objectType, object, fileCarrier, path.join(resDir, resName)]);
-
+        var javaParams = (req.path == '/hide') ?
+                        ['Java/src/steganography_tool/Steganography_Tool', 'Hide', objectType, object, fileCarrier, path.join(resDir, resName)] :
+                        ['Java/src/steganography_tool/Steganography_Tool', 'Retrieve', objectType, fileCarrier, path.join(resDir, resName)];
+        const ls = spawn('java', javaParams);
+        
         ls.stdout.on('data', (data) => {
-            console.log(`stdout: ${data}`);
+            if (req.path == '/retrieve') {
+                responseJSON["result"]+=`${data}`;
+            }
         });
 
         ls.stderr.on('data', (data) => {
@@ -200,13 +111,13 @@ router.post('/hide', function(req, res) {
         ls.on('close', (code) => {
             console.log('java process ended');
 
-            /*remove uploaded files*/
+            //remove uploaded files
             fs.unlink(fileCarrier, (err) => {
                 if (err) {
                     console.log("error deleting file carrier!");
                 }
             });
-            if (objectType == 'file') {
+            if ((objectType == 'file')&&(req.path=='/hide')) {
                 fs.unlink(object, (err) => {
                     if (err) {
                         console.log("error deleting object file!");
@@ -216,10 +127,12 @@ router.post('/hide', function(req, res) {
 
 
             if (responseJSON["status"] == 'success') {
-                responseJSON["result"] = resName;
-                setTimeout(() => {
-                    fs.unlink(path.resolve('./results/' + resName));
-                }, 7.2e6);
+                if ((objectType == 'file')&&(req.path=='/hide')) {
+                    responseJSON["result"] = resName;
+                    setTimeout(() => {
+                        fs.unlink(path.resolve('./results/' + resName));
+                    }, 7.2e6);
+                }
 
             }
             res.json(responseJSON);
@@ -231,6 +144,6 @@ router.post('/hide', function(req, res) {
 
     // parse the incoming request containing the form data
     form.parse(req);
-
 });
+
 module.exports = router;
